requirements.txt
airtable-python-wrapper
alembic==1.6.5
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
boto3==1.18.5
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
coverage==5.5
cryptography==3.4.7
dateparser==1.0.0
email_validator==1.1.3
flask-redis
freezegun==1.1.0
Flask==1.1.4
Flask-Babel==1.0.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-WTF==0.15.1
Flask-Cors==3.0.7
formfiller==0.3
gunicorn==20.1.0
idna==3.2
itsdangerous==1.1.0
# our 1.x Flask requires we stay at Jinja < 3.0
Jinja2==2.11.3
git+https://github.com/ty730/myvoteinfo
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
more-itertools==8.8.0
newrelic==6.4.4.161
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
pyusps==0.0.7
requests==2.26.0
robobrowser==0.5.3
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
usaddress==0.5.10
Werkzeug==1.0.1
WTForms==2.3.1
# WTForms >= 2.3.2 breaks validation of citizenship
# tests
behave==1.2.6
selenium==3.141.0
# For Gmail API
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
Wand==0.6.7
